fix: int -> long int (Sqlite count)

+++ include/sqlite_helper.h
@@ -1,6 +1,6 @@
 #ifndef SQLITE_HELPER_H
 #define SQLITE_HELPER_H
 
-int get_single_int(const char *filename_db, const char *query);
+long int get_single_int(const char *filename_db, const char *query);
 
 #endif
